+++ backend/api_business.py
@@ -23,7 +23,7 @@ from .datasources import (
     production_data,
 )
 from .erp import ErpConnectionError, ErpTimeoutError
-from .auth.authremote import validated_user
+from .auth.common import validated_user
 from .utils.responses import PdfStreamingResponse, ZipStreamingResponse
 from consolemsg import error
